updated readme, implenent event service and repo to rest of the table, tested all endpoint and working

+++ src/services/address-service.ts
@@ -1,5 +1,11 @@
 import { Address } from '../models/address';
 import { AddressRepository } from '../repos/address-repo';
+import { isValidId, 
+	isValidStrings, 
+	isValidObject, 
+	isPropertyOf, 
+	isEmptyObject } from '../util/validator'
+import { ResourceNotFoundError, BadRequestError } from '../errors/errors';
 
 export class AddressService{
 
@@ -10,28 +16,53 @@ export class AddressService{
 	async getAllAddress() : Promise<Address[]>{
 
 		let result = await this.addressRepo.getAll();
+		
+		if(isEmptyObject(result))
+			throw new ResourceNotFoundError()
+
 		return result;
 
 	}
+
 	async getAddressById(id: number) : Promise<Address> {
 
+		if(!isValidId(id))
+			throw new BadRequestError();
+
 		let result = await this.addressRepo.getById(id);
+
+		if(isEmptyObject(result))
+			throw new ResourceNotFoundError();
+		
 		return result;
 	}
 
 	async saveAddress(newObj: Address) : Promise<Address> {
+
+		if(!isValidObject(newObj, 'id'))
+			throw new BadRequestError('Invalid property values found in provided address.');
+
 		let result = await this.addressRepo.save(newObj);
+		
 		return result;
 	}
 
 	async updateAddress(updObj: Address) : Promise<boolean> {
-		await this.addressRepo.update(updObj);
-		return true;
+		if(!isValidObject(updObj))
+			throw new BadRequestError('Invalid address provided (invalid value found).');
+
+		let result = await this.addressRepo.update(updObj);
+
+		return result;
 	}
 
 	async deleteAddressById(id: number) : Promise<boolean> {
-		await this.addressRepo.deleteById(id);
-		return true;
+		if(!isValidId(id))
+			throw new BadRequestError();
 
+		let result = await this.addressRepo.deleteById(id);
+
+		return result;
 	}
+
 }
