+++ src/services/eventMember-service.ts
@@ -1,5 +1,11 @@
 import { EventMemberRepository } from '../repos/eventMember-repo';
 import { EventMember } from '../models/eventMember';
+import { isValidId, 
+	isValidStrings, 
+	isValidObject, 
+	isPropertyOf, 
+	isEmptyObject } from '../util/validator'
+import { ResourceNotFoundError, BadRequestError } from '../errors/errors';
 
 export class EventMemberService{
 
@@ -11,32 +17,43 @@ export class EventMemberService{
 
 		let result = await this.eventMemberRepo.getAll();
 
-		return result;
+		if(isEmptyObject(result))
+			throw new ResourceNotFoundError()
 
+		return result;
 	}
-	async getEventMemberById(id: number) : Promise<EventMember> {
+
+	async getEventMemberById(id: number) : Promise<EventMember[]> {
+
+		if(!isValidId(id))
+			throw new BadRequestError();
 
 		let result = await this.eventMemberRepo.getById(id);
+
+		if(isEmptyObject(result))
+			throw new ResourceNotFoundError();
 
 		return result;
 	}
 
 	async saveEventMember(newObj: EventMember) : Promise<EventMember> {
+
+		if(!isValidObject(newObj, 'id'))
+			throw new BadRequestError('Invalid property values found in provided events_members.');
+
 		let result = await this.eventMemberRepo.save(newObj);
             
 		return result;
 	}
 
-	async updateEventMember(updObj: EventMember) : Promise<boolean> {
-		await this.eventMemberRepo.update(updObj);
-		return true;
-	}
-
 	async deleteEventMemberById(id: number, id2: number) : Promise<boolean> {
 
-		await this.eventMemberRepo.deleteById(id, id2);
+		if(!isValidId(id))
+			throw new BadRequestError();
 
-		return true;
+		let result = await this.eventMemberRepo.deleteById(id, id2);
+
+		return result;
 
 	}
 }
