+++ src/services/member-service.ts
@@ -1,5 +1,11 @@
 import { Member } from '../models/member';
 import { MemberRepository } from '../repos/member-repo';
+import { isValidId, 
+	isValidStrings, 
+	isValidObject, 
+	isPropertyOf, 
+	isEmptyObject } from '../util/validator'
+import { ResourceNotFoundError, BadRequestError } from '../errors/errors';
 
 export class MemberService{
 
@@ -10,34 +16,53 @@ export class MemberService{
 	async getAllMember() : Promise<Member[]>{
 
 		let result = await this.memberRepo.getAll();
-		return result;
 
+		if(isEmptyObject(result))
+			throw new ResourceNotFoundError()
+
+		return result;
 	}
 	async getMemberById(id: number) : Promise<Member> {
 
-		let result = await this.memberRepo.getById(id);
-		return result;
+		if(!isValidId(id))
+			throw new BadRequestError();
 
+		let result = await this.memberRepo.getById(id);
+
+		if(isEmptyObject(result))
+			throw new ResourceNotFoundError();
+
+		return result;
 	}
 
 	async saveMember(newObj: Member) : Promise<Member> {
 
+		if(!isValidObject(newObj, 'id'))
+			throw new BadRequestError('Invalid property values found in provided member.');
+
 		let result = await this.memberRepo.save(newObj);
+
 		return result;
 	}
 
 	async updateMember(updObj: Member) : Promise<boolean> {
 
-		await this.memberRepo.update(updObj);
-		return true;
+		if(!isValidObject(updObj))
+			throw new BadRequestError('Invalid member provided (invalid value found).');
 
+		let result = await this.memberRepo.update(updObj);
+
+		return result;
 	}
 
 	async deleteMemberById(id: number) : Promise<boolean> {
 
-		await this.memberRepo.deleteById(id);
+		if(!isValidId(id))
+			throw new BadRequestError();
 
-		return true;
+		let result = await this.memberRepo.deleteById(id);
 
+		return result;
 	}
+
 }
